Create separate init binary
This removes the nsenter package from the main krood binary and
creates another binary that is used for starting containers.
