+++ cmd/kroo-init/init.go
@@ -1,0 +1,28 @@
+package main
+
+import (
+	"log"
+	"os"
+	"runtime"
+
+	"github.com/opencontainers/runc/libcontainer"
+	_ "github.com/opencontainers/runc/libcontainer/nsenter"
+)
+
+func init() {
+	if len(os.Args) > 1 && os.Args[1] == "init" {
+		runtime.GOMAXPROCS(1)
+		runtime.LockOSThread()
+		factory, err := libcontainer.New("")
+		if err != nil {
+			log.Fatal(err)
+		}
+
+		if err := factory.StartInitialization(); err != nil {
+			log.Fatal(err)
+		}
+		panic("--this line should have never been executed, congratulations--")
+	}
+}
+
+func main() {}
